Assessment #4/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
joblib>=1.1.0

# Visualization
plotly>=5.3.0
matplotlib>=3.4.3

# NLP and Text Processing
textblob>=0.15.3
nltk>=3.6.3

# Web Framework
flask>=2.0.1

# Communication
twilio>=7.0.0
pywhatkit>=5.3

# Testing
pytest>=6.2.5

# Utilities
python-dateutil>=2.8.2
requests>=2.26.0
threading2>=0.3.1

# Machine Learning
scikit-learn>=1.0.2
tensorflow>=2.8.0
keras>=2.8.0

# Data Processing
pandas-profiling>=3.2.0

# Monitoring and Analytics
prometheus-client>=0.13.0
grafana-api-client>=0.11.0